fix and enable diskfmt regress test for vmd(8)
Clean the rot in the diskfmt test, updating it to work with current
vmd(8) and vmctl(8). Remove the needs to qemu since vmctl can create
qcow2 images now.

Now that it's a test that doesn't need a port, enable it in the
parent Makefile.

There's still some work needed to get "make clean" properly cleaning
up after the tests.

OK ori@