add some devices found on the intel nuc10i7fnh
For the thunderbolt controller, while a public datasheet with product
ids and marketing names can't be found we know these ids are for the two
channel version of the thunderbolt 3 controller codenamed titan ridge
from public patches by Intel employees.  There are two channels per port
and the only single port titan ridge described on ark.intel.com is the
JHL7340.  The ids included with lspci refer to these devices as JHL7540
but that is a four channel / two port controller.

initial patch from fkr