.Fl marks up options not arguments
POSIX says `-v' is an option, not an argument;  options may take arguments,
as documented in getopt(3).

Programs may take arguments as well, e.g. "prog -v -- arg ...", but these
are usually marked up with `.Cm'.

Change "The arguments are as follows:" to "The options are as follows:" in
`.Fl's DESCRIPTION example to match what most of our base manuals say.

OK schwarze jmc