Optimize gadget fixups for MOV instructions.
Instead of swapping registers around, we can just use the REV version of
the same instruction, which has the same effect but encodes differently and
does not result in return bytes in the binary. This reduces the number
of xchg instructions resulting from gadget fixing.

Prompted by ratchov@, with input from millert@ and sthen@.

ok sthen@