make the printf on skipping reset with gen7/gen8 debug only
With the 5.10 drm on vlv/ivb/hsw/bdw inteldrm has been known to reset
the chip when a gpu hang is falsely detected.  This appears to be
related to the workarounds for cleaning up after switching hardware
contexts.

intel_gt_reset *NOTICE* [drm] Resetting chip for context closure in glsl-uniform-int<99177>
mark_guilty *NOTICE* [drm] glsl-uniform-int[97885] context reset due to GPU hang

A local change skips the reset (which kills Xorg) on gen7/gen8.
The printf is in that path.