Markup promises with Cm not Va
These are not variable names but rather keywords (as port of one big
string).

I originally changed this such that `.Bl -tag' would automatically tag them
and thus allow ":t id" in the manual pager;  since schwarze's recent
mandoc(1) commit this is no longer needed as `.Va' gained tagging with it,
but its use is incorrect here nonetheless, so still change it to the more
appropiate `.Cm'.

Input jmc schwarze