Fix automatic upgrade after fetching response file with dhcp.
When the auto upgrader tried to fetch the sets, sometimes ftp did
not find any.  The interface was down and had no address.  As
autoconf was not cleared, the upgrade's static netstart and dhcpleased
both tried to configure the interface, which cannot work.

Disable dhcp with "ifconfig inet -autoconf" in get_responsefile()
after it is not used anymore.

OK florian@